+++ backend/windmill-worker/src/python_executor.rs
@@ -888,6 +888,8 @@ pub async fn start_worker(
         None,
         None,
         None,
+        None,
+        None,
     )
     .await
     .to_vec();
@@ -1001,6 +1003,8 @@ for line in sys.stdin:
         Uuid::nil().to_string().as_str(),
         "dedicted_worker",
         Some(script_path.to_string()),
+        None,
+        None,
         None,
         None,
         None,